Rename parser.py to lman_parser.py to avoid name clash with deprecated stdlib module on Windows
